Fix showing available commands - correct path for subscription
Previously, there was no fixed place where Pykzee exposes information
on commands. It used to depend on where the user would place the
`core-plugin` in the state tree. So having a hard-coded path of
"/core/commands" was not right in the first place.

In the meantime, Pykzee was changed to always place the core plugin
in "/sys", which makes situations like this much easier: it is now
perfectly fine to hard-code the path. However, it happens to be
"/sys" rather than "/core", so this commit fixes this.